+++ src/interfaces/CursBeanR.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface CursBeanR {
     public void insertCursR(String nume, int numarStudenti, String numeProfesor);
-    public void updateCursR(Long id, String nume, int numarStudenti, String numeProfesor);
+    public void updateCursR(int id, String nume, int numarStudenti, String numeProfesor);
     public void deleteCursR(CursDTO cursDTO);
-    public CursDTO findCursR(Long id);
+    public CursDTO findCursR(int id);
     public List<CursDTO> findAllCsR();
-    public List<CursDTO> findAllCursForStR(Long idStudent);
+    public List<CursDTO> findAllCursForStR(int idStudent);
 }
